+++ src/components/Header.tsx
@@ -33,9 +33,9 @@ export const Header = () => {
   const isHistoric = flattenQuery(rideDate) < getNow();
 
   return (
-    <div className="fixed z-10 flex h-16 w-full items-center justify-center bg-neutral-100 text-neutral-700 sm:h-24">
+    <div className="fixed z-10 flex h-16 w-full items-center justify-center bg-blue-900  text-white sm:h-24 md:bg-neutral-100 md:text-neutral-700">
       <div className="container flex w-full flex-row justify-between px-2 md:px-4 lg:max-w-[1024px]">
-        <div className=" text-4xl font-bold tracking-wide sm:text-5xl">
+        <div className=" text-4xl tracking-wide sm:text-5xl">
           <button
             type="button"
             onClick={goHome}